important fix in ToFof, ToTff: goal at last position!
require that core modules have exactly one goal as last position in the module, test this and throw TransformationErrors if this is not the case!
(this is important because all constructs after a goal are not in scope anymore and hence should not become part of the translated fof/tff!